# task4final-project/swagger.yaml
swagger: "2.0"
info:
  description: "Trello like app"
  version: "1.0.0"
  title: "Trello like app"
  termsOfService: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
schemes:
- "https"
paths:
  /apitrello/project:
    post:
      tags: 
      - "project"
      description: ""
      operationId: "addProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New project description and name"
        required: true
        schema:
          $ref: "#/definitions/projectInit"
      responses:
        "200":
          description: "Created project"
          schema:
            $ref: "#/definitions/projectResult"
    get: 
      tags: 
      - "project"
      operationId: "getProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses: 
        "200":
          description: "All projects"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/projectResult"
    put:
      tags: 
      - "project"
      description: ""
      operationId: "putProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New project description and name"
        required: true
        schema:
          $ref: "#/definitions/projectResult"
      responses:
        "200":
          description: "Created project"
          schema:
            $ref: "#/definitions/projectResult"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  projectResult: 
    type: "object"
    properties:
      projectid:
        type: "integer"
      name:
        type: "string"
      description:
        type: "string"
...
  projectInit: 
    type: "object"
    required: 
    - "name"
    properties:
      name:
        type: "string"
      description:
        type: "string"
...